fix: extract recovery window to separate module to avoid crash loop
Move boot fail check to the top of createMainWindow() and delegate
entirely to recoveryWindow.ts when bootFailCount >= 3. This skips
all normal initialization (getMetadata, interceptFileProtocol, etc.)
that could throw and prevent createMainWindow from returning.

- recoveryWindow.ts: standalone module with its own BrowserWindow,
  IPC handlers, and uninterceptProtocol call
- app.ts: early return from createMainWindow, closed handler to
  null mainWindow, skip startServices/initMenu/initChildProcess
  in recovery mode
