+++ apps/desktop/app/recoveryWindow.ts
@@ -1,0 +1,167 @@
+import fs from 'fs';
+import path from 'path';
+import { format as formatUrl } from 'url';
+
+import { BrowserWindow, app, ipcMain, session } from 'electron';
+import logger from 'electron-log/main';
+
+import { ipcMessageKeys } from './config';
+import * as store from './libs/store';
+import { getAppStaticResourcesPath } from './resoucePath';
+
+const isMac = process.platform === 'darwin';
+
+/**
+ * Create a standalone recovery window with its own IPC handlers.
+ * This is completely independent from the normal createMainWindow flow —
+ * no getMetadata, no interceptFileProtocol, no bundle verification.
+ */
+export function createRecoveryWindow(): BrowserWindow {
+  const { screen } = require('electron');
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
+  const display = screen.getPrimaryDisplay();
+  const dimensions = display.workAreaSize;
+  const appStaticResourcesPath = getAppStaticResourcesPath();
+
+  // Clear any previously registered file protocol interceptor so
+  // recovery.html loads via the default file:// handler from the asar.
+  session.defaultSession.protocol.uninterceptProtocol('file');
+
+  const browserWindow = new BrowserWindow({
+    show: true,
+    title: 'OneKey',
+    titleBarStyle: 'hidden',
+    trafficLightPosition: { x: 20, y: 20 },
+    autoHideMenuBar: true,
+    frame: true,
+    resizable: true,
+    width: Math.min(800, dimensions.width),
+    height: Math.min(600, dimensions.height),
+    backgroundColor: '#0F0F0F',
+    webPreferences: {
+      spellcheck: false,
+      webviewTag: false,
+      webSecurity: true,
+      contextIsolation: false,
+      preload: path.join(__dirname, 'preload.js'),
+      sandbox: false,
+      nodeIntegration: false,
+    },
+    icon: path.join(appStaticResourcesPath, 'images/icons/512x512.png'),
+  });
+
+  const recoveryUrl = formatUrl({
+    pathname: path.join(__dirname, 'recovery.html'),
+    protocol: 'file',
+    slashes: true,
+  });
+  void browserWindow.loadURL(recoveryUrl);
+
+  // === IPC Handlers ===
+
+  ipcMain.removeHandler(ipcMessageKeys.RECOVERY_EXPORT_LOGS);
+  ipcMain.handle(ipcMessageKeys.RECOVERY_EXPORT_LOGS, async () => {
+    try {
+      const { dialog } = await import('electron');
+      const logDir = path.dirname(logger.transports.file.getFile().path);
+      const logFiles = fs.readdirSync(logDir).filter((f) => f.endsWith('.log'));
+      if (logFiles.length === 0) {
+        return { error: 'No log files found' };
+      }
+      const timestamp = new Date().toISOString().replace(/[:.]/g, '-');
+      const result = await dialog.showSaveDialog({
+        defaultPath: `onekey-logs-${timestamp}.zip`,
+        filters: [{ name: 'ZIP', extensions: ['zip'] }],
+      });
+      if (result.canceled || !result.filePath) {
+        return {};
+      }
+      const AdmZip = (await import('adm-zip')).default;
+      const zip = new AdmZip();
+      for (const file of logFiles) {
+        zip.addLocalFile(path.join(logDir, file));
+      }
+      zip.writeZip(result.filePath);
+      return {};
+    } catch (e: any) {
+      logger.error('Recovery export logs failed:', e);
+      return { error: e?.message || String(e) };
+    }
+  });
+
+  ipcMain.removeHandler(ipcMessageKeys.RECOVERY_TRY_AGAIN);
+  ipcMain.handle(ipcMessageKeys.RECOVERY_TRY_AGAIN, async () => {
+    store.resetConsecutiveBootFailCount();
+    if (process.mas) {
+      return { needsManualRestart: true };
+    }
+    app.relaunch();
+    app.exit(0);
+  });
+
+  ipcMain.removeHandler(ipcMessageKeys.RECOVERY_AUTO_REPAIR);
+  ipcMain.handle(ipcMessageKeys.RECOVERY_AUTO_REPAIR, async () => {
+    const errors: string[] = [];
+    try {
+      store.clearUpdateBundleData();
+    } catch (e: any) {
+      errors.push(`clearUpdateBundleData: ${e?.message}`);
+    }
+    try {
+      store.clearFallbackUpdateBundleData();
+    } catch (e: any) {
+      errors.push(`clearFallbackUpdateBundleData: ${e?.message}`);
+    }
+    try {
+      store.clearASCFile();
+    } catch (e: any) {
+      errors.push(`clearASCFile: ${e?.message}`);
+    }
+    try {
+      store.clearUpdateBuildNumber();
+    } catch (e: any) {
+      errors.push(`clearUpdateBuildNumber: ${e?.message}`);
+    }
+    try {
+      const bundleDir = path.join(app.getPath('userData'), 'onekey-bundle');
+      if (fs.existsSync(bundleDir)) {
+        fs.rmSync(bundleDir, { recursive: true, force: true });
+      }
+    } catch (e: any) {
+      errors.push(`deleteBundleDir: ${e?.message}`);
+    }
+    try {
+      const bundleDownloadDir = path.join(
+        app.getPath('userData'),
+        'onekey-bundle-download',
+      );
+      if (fs.existsSync(bundleDownloadDir)) {
+        fs.rmSync(bundleDownloadDir, { recursive: true, force: true });
+      }
+    } catch (e: any) {
+      errors.push(`deleteBundleDownloadDir: ${e?.message}`);
+    }
+    try {
+      store.clearMmkvRecoveryKeys();
+    } catch (e: any) {
+      errors.push(`clearMmkvKeys: ${e?.message}`);
+    }
+    store.resetConsecutiveBootFailCount();
+    if (errors.length > 0) {
+      logger.error('Recovery auto repair partial errors:', errors);
+    }
+    return errors.length > 0 ? { error: errors.join('; ') } : {};
+  });
+
+  // Mac: hide instead of close
+  // @ts-expect-error
+  browserWindow.on('close', (event: Event) => {
+    if (isMac) {
+      event.preventDefault();
+      browserWindow.blur();
+      browserWindow.hide();
+    }
+  });
+
+  return browserWindow;
+}
